requirements.txt
certifi==2019.11.28
cffi==1.13.2
cryptography==2.8
future==0.18.2
psycopg2==2.8.4
pycparser==2.19
python-telegram-bot==12.3.0
PyYAML==5.3
six==1.14.0
SQLAlchemy==1.3.12
tornado==6.0.3
